pyproject.toml
[project]
name = "saskia-mediapipe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "mediapipe>=0.10.32",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "ruff>=0.15.2",
]

[project.scripts]
pose-draw = "main:main"
